migrate `DefDatabase::macro_def`
