+++ crates/hir-def/src/db.rs
@@ -1,13 +1,11 @@
 //! Defines database & queries for name resolution.
 use base_db::{Crate, SourceDatabase};
-use hir_expand::{
-    EditionedFileId, HirFileId, InFile, Lookup, MacroCallId, MacroDefId, MacroDefKind,
-};
+use hir_expand::{EditionedFileId, HirFileId, MacroCallId};
 use salsa::{Durability, Setter};
 use triomphe::Arc;
 
 use crate::{
-    AssocItemId, AttrDefId, MacroExpander, MacroId, MacroRulesLocFlags, TraitId,
+    AssocItemId, AttrDefId, TraitId,
     attrs::AttrFlags,
     item_tree::{ItemTree, file_item_tree},
     nameres::crate_def_map,
@@ -20,10 +18,6 @@ pub trait DefDatabase: SourceDatabase {
     #[salsa::invoke(file_item_tree)]
     #[salsa::transparent]
     fn file_item_tree(&self, file_id: HirFileId, krate: Crate) -> &ItemTree;
-
-    /// Turns a MacroId into a MacroDefId, describing the macro's definition post name resolution.
-    #[salsa::invoke(macro_def)]
-    fn macro_def(&self, m: MacroId) -> MacroDefId;
 
     // region:visibilities
 
@@ -80,56 +74,4 @@ fn crate_supports_no_std(db: &dyn DefDatabase, crate_id: Crate) -> bool {
     let root_module = crate_def_map(db, crate_id).root_module_id();
     let attrs = AttrFlags::query(db, AttrDefId::ModuleId(root_module));
     attrs.contains(AttrFlags::IS_NO_STD)
-}
-
-fn macro_def(db: &dyn DefDatabase, id: MacroId) -> MacroDefId {
-    let kind = |expander, file_id, m| {
-        let in_file = InFile::new(file_id, m);
-        match expander {
-            MacroExpander::Declarative { styles } => MacroDefKind::Declarative(in_file, styles),
-            MacroExpander::BuiltIn(it) => MacroDefKind::BuiltIn(in_file, it),
-            MacroExpander::BuiltInAttr(it) => MacroDefKind::BuiltInAttr(in_file, it),
-            MacroExpander::BuiltInDerive(it) => MacroDefKind::BuiltInDerive(in_file, it),
-            MacroExpander::BuiltInEager(it) => MacroDefKind::BuiltInEager(in_file, it),
-            MacroExpander::UnimplementedBuiltIn => MacroDefKind::UnimplementedBuiltIn(in_file),
-        }
-    };
-
-    match id {
-        MacroId::Macro2Id(it) => {
-            let loc = it.lookup(db);
-
-            MacroDefId {
-                krate: loc.container.krate(db),
-                kind: kind(loc.expander, loc.id.file_id, loc.id.value.upcast()),
-                local_inner: false,
-                allow_internal_unsafe: loc.allow_internal_unsafe,
-                edition: loc.edition,
-            }
-        }
-        MacroId::MacroRulesId(it) => {
-            let loc = it.lookup(db);
-
-            MacroDefId {
-                krate: loc.container.krate(db),
-                kind: kind(loc.expander, loc.id.file_id, loc.id.value.upcast()),
-                local_inner: loc.flags.contains(MacroRulesLocFlags::LOCAL_INNER),
-                allow_internal_unsafe: loc
-                    .flags
-                    .contains(MacroRulesLocFlags::ALLOW_INTERNAL_UNSAFE),
-                edition: loc.edition,
-            }
-        }
-        MacroId::ProcMacroId(it) => {
-            let loc = it.lookup(db);
-
-            MacroDefId {
-                krate: loc.container.krate(db),
-                kind: MacroDefKind::ProcMacro(loc.id, loc.expander, loc.kind),
-                local_inner: false,
-                allow_internal_unsafe: false,
-                edition: loc.edition,
-            }
-        }
-    }
 }
